MNT: Move lxe tests to test_lxe.py
